+++ script/History.markdown
@@ -1,3 +1,11 @@
+### 0.3.0 / 2012-10-07
+
+* Add --delete option to delete all records
+
+### 0.2.0 / 2012-10-06
+
+* Add --load option for builtin fixtures
+
 ### 0.1.0 / 2012-10-06
 
 * Everything is new. First release